+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Breaking changes should be prefixed by `[**BREAKING**]` (without the quotes), to
 - Announce (boost/RT), like or dislike `Fedipub::DataEntity` and `Fedipub::Actor` objects using new `announce!`, `like!` and `dislike!` methods.
 - Reverse any `Fedipub::Activity` by calling `undo!` to create an Undo activity addressed in the same way.
 - Set `results` and `instrument` fields for outgoing `Fedipub::Activity` objects.
+- Reduce schema lookups to W3C security vocabulary by preloading context.
 
 ### Fixed